[inf lab3] got 80% of the grade
Apparently I'd overestimated the soundness of my algos. At least
one of them is outright wrong (Shannon-Fano, if you're wondering --
the branch splitting part, to be precise. It needs to find lhs and rhs
so that lhs has higher sum of probabilities _and_ the difference
between sums of probabilities of lhs and rhs is the least possible;
my implementation only guarantees the former to be true.)

That, and I hadn't completely answered another question about
information entropy. Perhaps I should have paid more attention
to the theoretical material... Now that I think about it, nah.
I had a lot of fun doing this, after all, and that's what really matters.