[dm cw] I admit my defeat
I don't have enough knowledge to automate this at the moment. The
hardest part by far is presentation, there's a well-defined format
we are expected to follow, and reproducing it in LaTeX with
dynamically generated content has proven to be such a... hassle,
I do not think it's worth it.